text_analysis_platform/requirements.txt
streamlit
plotly
matplotlib
pandas
numpy
spacy
gensim
wordcloud
python-docx
scikit-learn
nltk
transformers
torch